requirements.txt
# ── Core RAG stack ──────────────────────────────────────────────
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
langchain-huggingface==0.0.3
langchain-chroma==0.1.3

# ── Vector store ─────────────────────────────────────────────────
chromadb==0.5.3

# ── Embeddings (free, runs locally) ──────────────────────────────
sentence-transformers==3.0.1

# ── Document loaders ─────────────────────────────────────────────
pypdf==4.3.1
python-docx==1.1.2
unstructured==0.15.0

# ── LLM provider (free tier) ─────────────────────────────────────
groq==0.9.0

# ── Re-ranking ───────────────────────────────────────────────────
sentence-transformers==3.0.1   # reused for cross-encoder

# ── Evaluation ───────────────────────────────────────────────────
ragas==0.1.14
datasets==2.20.0

# ── API backend ──────────────────────────────────────────────────
fastapi==0.112.0
uvicorn==0.30.5
python-multipart==0.0.9

# ── UI ───────────────────────────────────────────────────────────
streamlit==1.37.0

# ── Utilities ────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.8.2
loguru==0.7.2
pandas==2.2.2
